Change blurb.
`heks` gets at the core of what bothers me in many applications: user
interaction. That is why it should be my priority for now.

If done correctly, it could subsume a lot of the compulsions I feel for
writing other software. For example: I could get a lot of the behaviour
I'd want out of `stilish` by simply combining `heks` with `tmux`.

(Yes, I'll get to coding someday.)